refactor: replace customResponse with baseResponse
