+++ src/main/java/com/keepgoing/keepserver/domain/book/presentation/BookController.java
@@ -2,6 +2,7 @@ package com.keepgoing.keepserver.domain.book.presentation;
 
 import com.keepgoing.keepserver.domain.book.entity.Book;
 import com.keepgoing.keepserver.domain.book.repository.BookRepository;
+import com.keepgoing.keepserver.global.common.BaseResponse;
 import com.keepgoing.keepserver.global.dto.response.CustomResponseEntity;
 import com.keepgoing.keepserver.domain.book.service.BookService;
 import lombok.RequiredArgsConstructor;
@@ -21,38 +22,21 @@ import java.util.List;
 @RequestMapping("/book")
 public class BookController {
     private static final Logger logger = LoggerFactory.getLogger(BookController.class);
-    private final BookRepository bookRepository;
     private final BookService bookService;
 
-    @PostMapping("/register2")
-    public List<Book> allBook(){
-        return bookService.selectAllBook();
-    }
-
-
-
     @PostMapping("/register")
-    public CustomResponseEntity<String> bookRegistration(@RequestPart(value = "book") Book book,
-                                                         @RequestPart(value = "image") MultipartFile multipartFile) {
-        bookService.bookRegistration(book, multipartFile);
-        return new CustomResponseEntity<String>(
-                HttpStatus.OK
-                , "BookAdding Successful"
-        );
+    public BaseResponse bookRegistration(@RequestPart(value = "book") Book book,
+                                         @RequestPart(value = "image") MultipartFile multipartFile) {
+        return bookService.bookRegister(book,multipartFile);
     }
 
     @GetMapping("/all")
-    public CustomResponseEntity<ArrayList<Book>> selectAllBook() {
-        return new CustomResponseEntity<ArrayList<Book>>(
-                HttpStatus.OK
-                , (ArrayList<Book>) bookService.selectAllBook()
-        );
+    public BaseResponse selectAllBook() {
+        return bookService.selectAllBook();
     }
 
     @DeleteMapping("/del/{nfcCode}")
-    public CustomResponseEntity<String> deleteBookByNfcCode(@PathVariable(value = "nfcCode") String nfcCode) {
-        return new CustomResponseEntity<String>(HttpStatus.OK
-                , bookService.deleteBookByNfcCode(nfcCode)
-        );
+    public BaseResponse deleteBookByNfcCode(@PathVariable(value = "nfcCode") String nfcCode) {
+        return bookService.deleteBook(nfcCode);
     }
 }
