+++ src/main/java/com/keepgoing/keepserver/domain/book/service/BookServiceImpl.java
@@ -3,9 +3,10 @@ package com.keepgoing.keepserver.domain.book.service;
 import com.keepgoing.keepserver.domain.book.entity.Book;
 import com.keepgoing.keepserver.domain.book.repository.BookRepository;
 import com.keepgoing.keepserver.domain.book.util.GenerateCertCharacter;
-import com.keepgoing.keepserver.global.dto.response.BaseResponse;
+import com.keepgoing.keepserver.domain.device.repository.DeviceRepository;
+import com.keepgoing.keepserver.global.common.BaseResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -17,34 +18,35 @@ import java.util.List;
 @Service
 public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
+    private final DeviceRepository deviceRepository;
 
     @Override
-    public void bookRegistration(Book book, MultipartFile multipartFile) {
+    public BaseResponse bookRegister(Book book, MultipartFile multipartFile){
         String nfcCode = createNfcCode();
         book.setRegistrationDate(new Date());
         book.setNfcCode(nfcCode);
         book.setState("0");
         bookRepository.save(book);
+        return new BaseResponse(HttpStatus.OK, "책 생성 성공");
     }
 
 
     @Override
-    public ArrayList<Book> selectAllBook() {
-        return (ArrayList<Book>) bookRepository.findAll();
-
+    public BaseResponse selectAllBook() {
+        return new BaseResponse(HttpStatus.OK, "책 불러오기 성공",(ArrayList<Book>) bookRepository.findAll());
     }
 
     @Override
-    public String deleteBookByNfcCode(String nfcCode) {
+    public BaseResponse deleteBook(String nfcCode) {
         if (nfcCode == null || nfcCode.isEmpty()) {
-            return "Invalid NFC code";
+            return new BaseResponse(HttpStatus.FORBIDDEN,"Invalid NFC code");
         }
         Book book = bookRepository.findBookByNfcCode(nfcCode);
         if (book == null) {
-            return "This book cannot be found";
+            return new BaseResponse(HttpStatus.OK,"This book cannot be found");
         } else {
             bookRepository.delete(bookRepository.findBookByNfcCode(nfcCode));
-            return "Deletion successful";
+            return new BaseResponse(HttpStatus.OK,"Deletion successful");
         }
     }
 
@@ -57,13 +59,6 @@ public class BookServiceImpl implements BookService {
         } while (bookRepository.findBookByNfcCode(newNfcCode) != null);
         return newNfcCode;
     }
-
-    @Override
-    public BaseResponse findAll() {
-        List<Book> books = bookRepository.findAll();
-        return null;
-    }
-
 
 }
